chore: refactor cibuildwheel.utils
