+++ .gitignore
@@ -84,9 +84,7 @@ celerybeat-schedule
 
 # virtualenv
 .venv
-venv/
-venv3/
-venv2/
+venv*/
 ENV/
 env/
 env2/
@@ -111,9 +109,6 @@ all_known_setup.yaml
 
 # mkdocs
 site/
-
-# Virtual environments
-venv*
 
 # PyCharm
 .idea/
