+++ cibuildwheel/pyodide.py
@@ -1,39 +1,41 @@
 from __future__ import annotations
 
+import functools
 import os
 import shutil
 import sys
+import tomllib
 from collections.abc import Sequence, Set
 from dataclasses import dataclass
 from pathlib import Path
+from tempfile import TemporaryDirectory
+from typing import Final
 
 from filelock import FileLock
 
 from . import errors
 from .architecture import Architecture
 from .environment import ParsedEnvironment
+from .frontend import BuildFrontendConfig, get_build_verbosity_extra_flags, split_config_settings
 from .logger import log
 from .options import Options
+from .selector import BuildSelector
 from .typing import PathOrStr
-from .util import (
+from .util import resources
+from .util.cmd import call, shell
+from .util.file import (
     CIBW_CACHE_PATH,
-    BuildFrontendConfig,
-    BuildSelector,
-    call,
-    combine_constraints,
     copy_test_sources,
     download,
-    ensure_node,
+    extract_tar,
     extract_zip,
-    find_compatible_wheel,
-    get_pip_version,
     move_file,
-    prepare_command,
-    read_python_configs,
-    shell,
-    split_config_settings,
-    virtualenv,
 )
+from .util.helpers import prepare_command
+from .util.packaging import combine_constraints, find_compatible_wheel, get_pip_version
+from .venv import virtualenv
+
+IS_WIN: Final[bool] = sys.platform.startswith("win")
 
 
 @dataclass(frozen=True)
@@ -44,6 +46,37 @@ class PythonConfiguration:
     pyodide_build_version: str
     emscripten_version: str
     node_version: str
+
+
+@functools.cache
+def ensure_node(major_version: str) -> Path:
+    with resources.NODEJS.open("rb") as f:
+        loaded_file = tomllib.load(f)
+    version = str(loaded_file[major_version])
+    base_url = str(loaded_file["url"])
+    ext = "zip" if IS_WIN else "tar.xz"
+    platform = "win" if IS_WIN else ("darwin" if sys.platform.startswith("darwin") else "linux")
+    linux_arch = Architecture.native_arch("linux")
+    assert linux_arch is not None
+    arch = {"x86_64": "x64", "i686": "x86", "aarch64": "arm64"}.get(
+        linux_arch.value, linux_arch.value
+    )
+    name = f"node-{version}-{platform}-{arch}"
+    path = CIBW_CACHE_PATH / name
+    with FileLock(str(path) + ".lock"):
+        if not path.exists():
+            url = f"{base_url}{version}/{name}.{ext}"
+            with TemporaryDirectory() as tmp_path:
+                archive = Path(tmp_path) / f"{name}.{ext}"
+                download(url, archive)
+                if ext == "zip":
+                    extract_zip(archive, path.parent)
+                else:
+                    extract_tar(archive, path.parent)
+    assert path.exists()
+    if not IS_WIN:
+        return path / "bin"
+    return path
 
 
 def install_emscripten(tmp: Path, version: str) -> Path:
@@ -188,7 +221,7 @@ def get_python_configurations(
     build_selector: BuildSelector,
     architectures: Set[Architecture],  # noqa: ARG001
 ) -> list[PythonConfiguration]:
-    full_python_configs = read_python_configs("pyodide")
+    full_python_configs = resources.read_python_configs("pyodide")
 
     python_configurations = [PythonConfiguration(**item) for item in full_python_configs]
     python_configurations = [c for c in python_configurations if build_selector(c.identifier)]
@@ -285,10 +318,9 @@ def build(options: Options, tmp_path: Path) -> None:
 
                 extra_flags = split_config_settings(build_options.config_settings, "build")
                 extra_flags += build_frontend.args
-
-                if not 0 <= build_options.build_verbosity < 2:
-                    msg = f"build_verbosity {build_options.build_verbosity} is not supported for build frontend. Ignoring."
-                    log.warning(msg)
+                extra_flags = get_build_verbosity_extra_flags(
+                    build_options.build_verbosity, "build"
+                )
 
                 build_env = env.copy()
                 if build_options.dependency_constraints:
