+++ unit_test/download_test.py
@@ -5,7 +5,7 @@ import ssl
 import certifi
 import pytest
 
-from cibuildwheel.util import download
+from cibuildwheel.util.file import download
 
 DOWNLOAD_URL = "https://raw.githubusercontent.com/pypa/cibuildwheel/v1.6.3/requirements-dev.txt"
