+++ unit_test/validate_schema_test.py
@@ -7,12 +7,13 @@ from pathlib import Path
 import pytest
 import validate_pyproject.api
 
+from cibuildwheel.util import resources
+
 DIR = Path(__file__).parent.resolve()
 
 
 def test_validate_default_schema():
-    filepath = DIR.parent / "cibuildwheel/resources/defaults.toml"
-    with filepath.open("rb") as f:
+    with resources.DEFAULTS.open("rb") as f:
         example = tomllib.load(f)
 
     validator = validate_pyproject.api.Validator()
